requirements.txt
Django==1.11.3
psycopg2==2.6.2
djangorestframework
markdown
django-filter
django-cors-headers